Correctly pluralize floats between 1 and 2
This fixes https://github.com/stefanpenner/ember-inflector/issues/94

It also adds some test additional test cases to document the current
behavior.

One additional strange bit is that the float 1.0 is returned as 1 when
concatenated into the string. I'm not sure if this is the ideal response
but I did not see any clear way to handle keeping that as 1.0 in the
returned string. This is the current behavior so changing it might
cause some unintended side-effects.